requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.112.0
flatbuffers==24.3.25
fonttools==4.55.1
h11==0.14.0
idna==3.10
jax==0.4.35
jaxlib==0.4.35
kiwisolver==1.4.7
matplotlib==3.8.4
mediapipe==0.10.14
ml_dtypes==0.5.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.8.0.76
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.5